refactor: moves static json fn to fixtures
